refactor OEF search docker image
