+++ tests/conftest.py
@@ -18,7 +18,6 @@
 # ------------------------------------------------------------------------------
 
 """Conftest module for Pytest."""
-import asyncio
 import inspect
 import logging
 import os
@@ -31,7 +30,6 @@ import threading
 import time
 from functools import WRAPPER_ASSIGNMENTS, wraps
 from pathlib import Path
-from threading import Timer
 from types import FunctionType, MethodType
 from typing import Callable, List, Optional, Sequence, cast
 from unittest.mock import patch
@@ -39,8 +37,6 @@ from unittest.mock import patch
 import docker as docker
 import gym
 import pytest
-from docker.models.containers import Container
-from oef.agents import AsyncioCore, OEFAgent
 
 from aea import AEA_DIR
 from aea.aea import AEA
@@ -83,6 +79,7 @@ from packages.fetchai.connections.stub.connection import StubConnection
 from packages.fetchai.connections.tcp.tcp_client import TCPClientConnection
 from packages.fetchai.connections.tcp.tcp_server import TCPServerConnection
 
+from tests.common.docker_image import DockerImage, OEFSearchDockerImage
 from tests.data.dummy_connection.connection import DummyConnection  # type: ignore
 
 
@@ -460,144 +457,6 @@ def wait_for_localhost_ports_to_close(
         raise ValueError("Some ports are open: {}!".format(open_ports))
 
 
-class OEFHealthCheck(object):
-    """A health check class."""
-
-    def __init__(
-        self,
-        oef_addr: str,
-        oef_port: int,
-        loop: Optional[asyncio.AbstractEventLoop] = None,
-    ):
-        """
-        Initialize.
-
-        :param oef_addr: IP address of the OEF node.
-        :param oef_port: Port of the OEF node.
-        """
-        self.oef_addr = oef_addr
-        self.oef_port = oef_port
-
-        self._result = False
-        self._stop = False
-        self._core = AsyncioCore()
-        self.agent = OEFAgent(
-            "check", core=self._core, oef_addr=self.oef_addr, oef_port=self.oef_port
-        )
-        self.agent.on_connect_success = self.on_connect_ok
-        self.agent.on_connection_terminated = self.on_connect_terminated
-        self.agent.on_connect_failed = self.exception_handler
-
-    def exception_handler(self, url=None, ex=None):
-        """Handle exception during a connection attempt."""
-        print("An error occurred. Exception: {}".format(ex))
-        self._stop = True
-
-    def on_connect_ok(self, url=None):
-        """Handle a successful connection."""
-        print("Connection OK!")
-        self._result = True
-        self._stop = True
-
-    def on_connect_terminated(self, url=None):
-        """Handle a connection failure."""
-        print("Connection terminated.")
-        self._stop = True
-
-    def run(self) -> bool:
-        """
-        Run the check, asynchronously.
-
-        :return: True if the check is successful, False otherwise.
-        """
-        self._result = False
-        self._stop = False
-
-        def stop_connection_attempt(self):
-            if self.agent.state == "connecting":
-                self.agent.state = "failed"
-
-        t = Timer(1.5, stop_connection_attempt, args=(self,))
-
-        try:
-            print("Connecting to {}:{}...".format(self.oef_addr, self.oef_port))
-            self._core.run_threaded()
-
-            t.start()
-            self._result = self.agent.connect()
-            self._stop = True
-
-            if self._result:
-                print("Connection established. Tearing down connection...")
-                self.agent.disconnect()
-                t.cancel()
-            else:
-                print("A problem occurred. Exiting...")
-            return self._result
-
-        except Exception as e:
-            print(str(e))
-            return self._result
-        finally:
-            t.join(1.0)
-            self.agent.stop()
-            self.agent.disconnect()
-            self._core.stop()
-
-
-def _stop_oef_search_images():
-    """Stop the OEF search image."""
-    client = docker.from_env()
-    for container in client.containers.list():
-        if "fetchai/oef-search:0.7" in container.image.tags:
-            logger.info("Stopping existing Docker image...")
-            container.stop()
-
-
-def _wait_for_oef(max_attempts: int = 15, sleep_rate: float = 1.0):
-    """Wait until the OEF is up."""
-    success = False
-    attempt = 0
-    while not success and attempt < max_attempts:
-        attempt += 1
-        logger.info("Attempt {}...".format(attempt))
-        oef_healthcheck = OEFHealthCheck("127.0.0.1", 10000)
-        result = oef_healthcheck.run()
-        if result:
-            success = True
-        else:
-            logger.info(
-                "OEF not available yet - sleeping for {} second...".format(sleep_rate)
-            )
-            time.sleep(sleep_rate)
-
-    return success
-
-
-def _create_oef_docker_image(oef_addr_, oef_port_) -> Container:
-    client = docker.from_env()
-
-    logger.info(ROOT_DIR + "/tests/common/oef_search_pluto_scripts")
-    ports = {
-        "20000/tcp": ("0.0.0.0", 20000),  # nosec
-        "30000/tcp": ("0.0.0.0", 30000),  # nosec
-        "{}/tcp".format(oef_port_): ("0.0.0.0", oef_port_),  # nosec
-    }
-    volumes = {
-        ROOT_DIR
-        + "/tests/common/oef_search_pluto_scripts": {"bind": "/config", "mode": "rw"},
-        ROOT_DIR + "/data/oef-logs": {"bind": "/logs", "mode": "rw"},
-    }
-    c = client.containers.run(
-        "fetchai/oef-search:0.7",
-        "/config/node_config.json",
-        detach=True,
-        ports=ports,
-        volumes=volumes,
-    )
-    return c
-
-
 def pytest_addoption(parser) -> None:
     """Add --aea-loop option."""
     parser.addoption(
@@ -653,32 +512,35 @@ def network_node(
     oef_addr, oef_port, pytestconfig, timeout: float = 2.0, max_attempts: int = 10
 ):
     """Network node initialization."""
-    if sys.version_info < (3, 7):
-        pytest.skip("Python version < 3.7 not supported by the OEF.")
-        return
+    client = docker.from_env()
+    image = OEFSearchDockerImage(client, oef_addr, oef_port)
+    yield from _launch_image(image, timeout, max_attempts)
 
-    if os.name == "nt":
-        pytest.skip("Skip test as it doesn't work on Windows.")
 
-    _stop_oef_search_images()
-    c = _create_oef_docker_image(oef_addr, oef_port)
-    c.start()
+def _launch_image(image: DockerImage, timeout: float = 2.0, max_attempts: int = 10):
+    """
+    Launch image.
 
-    # wait for the setup...
-    logger.info("Setting up the OEF node...")
-    success = _wait_for_oef(max_attempts=max_attempts, sleep_rate=timeout)
-
+    :param image: an instancoe of Docker image.
+    :return: None
+    """
+    image.check_skip()
+    image.stop_if_already_running()
+    container = image.create()
+    container.start()
+    logger.info(f"Setting up image '%s'...", image.tag)
+    success = image.wait(max_attempts, timeout)
     if not success:
-        c.stop()
-        c.remove()
-        pytest.fail("OEF doesn't work. Exiting...")
+        container.stop()
+        container.remove()
+        pytest.fail(f"{image.tag} doesn't work. Exiting...")
     else:
         logger.info("Done!")
         time.sleep(timeout)
         yield
-        logger.info("Stopping the OEF node...")
-        c.stop()
-        c.remove()
+        logger.info("Stopping the image '%s'...", image.tag)
+        container.stop()
+        container.remove()
 
 
 @pytest.fixture(scope="session", autouse=True)
